ai-extraction-system/requirements.txt
openai>=1.0.0
flask>=2.0.0
flask-cors>=4.0.0
jsonschema>=4.0.0
tiktoken>=0.5.0
python-dotenv>=1.0.0
requests>=2.28.0
